backend/pyproject.toml
[project]
name = "attention-map-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.0",
    "django-stubs-ext>=5.2.7",
    "groq>=0.36.0",
    "gunicorn>=23.0.0",
    "minio>=7.2.20",
    "openai>=2.8.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "ruff>=0.14.7",
    "pytest-xdist>=3.8.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "attention_map.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-v -n auto --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["core", "api", "tasks", "services"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "PLC0415", # import-outside-toplevel
]
